pyproject.toml
[project]
name = "showcase-site"
version = "0.1.0"
requires-python = ">= 3.11"
dependencies = [
    "Flask==3.0.2",
    "numpy==1.26.4",
    "pandas==2.2.1",
    "python-dotenv==1.0.1",
    "SQLAlchemy==2.0.28",
    "schedule==1.2.1",
    "requests==2.31.0",
    "psycopg2==2.9.9",
    "flask-smorest==0.44.0",
    "lxml==5.1.0",
    "marshmallow_sqlalchemy=1.0.0",
]